% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02}
\alias{day02}
\alias{f02a}
\alias{f02b}
\alias{example_data_02}
\title{Day 02: Rock Paper Scissors}
\usage{
f02a(path)

f02b(x)

example_data_02(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f02a(x)} returns .... For Part Two,
\code{f02b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/2}{Rock Paper Scissors}
}
\details{
\strong{Part One}

The Elves begin to set up camp on the beach. To decide whose tent gets
to be closest to the snack storage, a giant \href{https://en.wikipedia.org/wiki/Rock_paper_scissors}{Rock Paper Scissors}{target="_blank"}
tournament is already in progress.

Rock Paper Scissors is a game between two players. Each game contains
many rounds; in each round, the players each simultaneously choose one
of Rock, Paper, or Scissors using a hand shape. Then, a winner for that
round is selected: Rock defeats Scissors, Scissors defeats Paper, and
Paper defeats Rock. If both players choose the same shape, the round
instead ends in a draw.

Appreciative of your help yesterday, one Elf gives you an \emph{encrypted
strategy guide} (your puzzle input) that they say will be sure to help
you win. \"The first column is what your opponent is going to play: \code{A}
for Rock, \code{B} for Paper, and \code{C} for Scissors. The second column\--\"
Suddenly, the Elf is called away to help with someone\'s tent.

The second column, [you reason]{title="Why do you keep guessing?!"},
must be what you should play in response: \code{X} for Rock, \code{Y} for Paper,
and \code{Z} for Scissors. Winning every time would be suspicious, so the
responses must have been carefully chosen.

The winner of the whole tournament is the player with the highest score.
Your \emph{total score} is the sum of your scores for each round. The score
for a single round is the score for the \emph{shape you selected} (1 for
Rock, 2 for Paper, and 3 for Scissors) plus the score for the \emph{outcome
of the round} (0 if you lost, 3 if the round was a draw, and 6 if you
won).

Since you can\'t be sure if the Elf is trying to help you or trick you,
you should calculate the score you would get if you were to follow the
strategy guide.

For example, suppose you were given the following strategy guide:

\if{html}{\out{<div class="sourceCode">}}\preformatted{A Y
B X
C Z
}\if{html}{\out{</div>}}

This strategy guide predicts and recommends the following:
\itemize{
\item In the first round, your opponent will choose Rock (\code{A}), and you
should choose Paper (\code{Y}). This ends in a win for you with a score
of \emph{8} (2 because you chose Paper + 6 because you won).
\item In the second round, your opponent will choose Paper (\code{B}), and you
should choose Rock (\code{X}). This ends in a loss for you with a score
of \emph{1} (1 + 0).
\item The third round is a draw with both players choosing Scissors,
giving you a score of 3 + 3 = \emph{6}.
}

In this example, if you were to follow the strategy guide, you would get
a total score of \emph{\code{15}} (8 + 1 + 6).

\emph{What would your total score be if everything goes exactly according to
your strategy guide?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f02a(example_data_02())
f02b()
}
