% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{f03a}
\alias{f03b}
\alias{example_data_03}
\title{Day 03: Rucksack Reorganization}
\usage{
f03a(input)

f03b(input)

example_data_03(example = 1)
}
\arguments{
\item{example}{Which example data to use (by position or name). Defaults to
1.}

\item{x}{some data}
}
\value{
For Part One, \code{f03a(x)} returns .... For Part Two,
\code{f03b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/3}{Rucksack Reorganization}
}
\details{
\strong{Part One}

One Elf has the important job of loading all of the
\href{https://en.wikipedia.org/wiki/Rucksack}{rucksacks}{target="_blank"}
with supplies for the
[jungle]{title="Where there's jungle, there's hijinxs."} journey.
Unfortunately, that Elf didn\'t quite follow the packing instructions,
and so a few items now need to be rearranged.

Each rucksack has two large \emph{compartments}. All items of a given type
are meant to go into exactly one of the two compartments. The Elf that
did the packing failed to follow this rule for exactly one item type per
rucksack.

The Elves have made a list of all of the items currently in each
rucksack (your puzzle input), but they need your help finding the
errors. Every item type is identified by a single lowercase or uppercase
letter (that is, \code{a} and \code{A} refer to different types of items).

The list of items for each rucksack is given as characters all on a
single line. A given rucksack always has the same number of items in
each of its two compartments, so the first half of the characters
represent items in the first compartment, while the second half of the
characters represent items in the second compartment.

For example, suppose you have the following list of contents from six
rucksacks:

\if{html}{\out{<div class="sourceCode">}}\preformatted{vJrwpWtwJgWrhcsFMMfFFhFp
jqHRNqRjqzjGDLGLrsFMfFZSrLrFZsSL
PmmdzqPrVvPwwTWBwg
wMqvLMZHhHMvwLHjbvcjnnSBnvTQFn
ttgJtRGJQctTZtZT
CrZsJsPPZsGzwwsLwLmpwMDw
}\if{html}{\out{</div>}}
\itemize{
\item The first rucksack contains the items \code{vJrwpWtwJgWrhcsFMMfFFhFp},
which means its first compartment contains the items \code{vJrwpWtwJgWr},
while the second compartment contains the items \code{hcsFMMfFFhFp}. The
only item type that appears in both compartments is lowercase \emph{\code{p}}.
\item The second rucksack\'s compartments contain \code{jqHRNqRjqzjGDLGL} and
\code{rsFMfFZSrLrFZsSL}. The only item type that appears in both
compartments is uppercase \emph{\code{L}}.
\item The third rucksack\'s compartments contain \code{PmmdzqPrV} and
\code{vPwwTWBwg}; the only common item type is uppercase \emph{\code{P}}.
\item The fourth rucksack\'s compartments only share item type \emph{\code{v}}.
\item The fifth rucksack\'s compartments only share item type \emph{\code{t}}.
\item The sixth rucksack\'s compartments only share item type \emph{\code{s}}.
}

To help prioritize item rearrangement, every item type can be converted
to a \emph{priority}:
\itemize{
\item Lowercase item types \code{a} through \code{z} have priorities 1 through 26.
\item Uppercase item types \code{A} through \code{Z} have priorities 27 through 52.
}

In the above example, the priority of the item type that appears in both
compartments of each rucksack is 16 (\code{p}), 38 (\code{L}), 42 (\code{P}), 22 (\code{v}),
20 (\code{t}), and 19 (\code{s}); the sum of these is \emph{\code{157}}.

Find the item type that appears in both compartments of each rucksack.
\emph{What is the sum of the priorities of those item types?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f03a(example_data_03())
f03b()
}
