% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{day07}
\alias{day07}
\alias{f07a}
\alias{f07b}
\alias{example_data_07}
\title{Day 07: No Space Left On Device}
\usage{
f07a(x)

f07b(x)

example_data_07(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f07a(x)} returns .... For Part Two,
\code{f07b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2022/day/7}{No Space Left On Device}
}
\details{
\strong{Part One}

You can hear birds chirping and raindrops hitting leaves as the
expedition proceeds. Occasionally, you can even hear much louder sounds
in the distance; how big do the animals get out here, anyway?

The device the Elves gave you has problems with more than just its
communication system. You try to run a system update:

\if{html}{\out{<div class="sourceCode">}}\preformatted{$ system-update --please --pretty-please-with-sugar-on-top
Error: No space left on device
}\if{html}{\out{</div>}}

Perhaps you can delete some files to make space for the update?

You browse around the filesystem to assess the situation and save the
resulting terminal output (your puzzle input). For example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{$ cd /
$ ls
dir a
14848514 b.txt
8504156 c.dat
dir d
$ cd a
$ ls
dir e
29116 f
2557 g
62596 h.lst
$ cd e
$ ls
584 i
$ cd ..
$ cd ..
$ cd d
$ ls
4060174 j
8033020 d.log
5626152 d.ext
7214296 k
}\if{html}{\out{</div>}}

The filesystem consists of a tree of files (plain data) and directories
(which can contain other directories or files). The outermost directory
is called \code{/}. You can navigate around the filesystem, moving into or
out of directories and listing the contents of the directory you\'re
currently in.

Within the terminal output, lines that begin with \code{$} are \emph{commands you
executed}, very much like some modern computers:
\itemize{
\item \code{cd} means \emph{change directory}. This changes which directory is the
current directory, but the specific result depends on the argument:
\itemize{
\item \verb{cd x} moves \emph{in} one level: it looks in the current directory
for the directory named \code{x} and makes it the current directory.
\item \verb{cd ..} moves \emph{out} one level: it finds the directory that
contains the current directory, then makes that directory the
current directory.
\item \verb{cd /} switches the current directory to the outermost
directory, \code{/}.
}
\item \code{ls} means \emph{list}. It prints out all of the files and directories
immediately contained by the current directory:
\itemize{
\item \verb{123 abc} means that the current directory contains a file named
\code{abc} with size \code{123}.
\item \verb{dir xyz} means that the current directory contains a directory
named \code{xyz}.
}
}

Given the commands and output in the example above, you can determine
that the filesystem looks visually like this:

\if{html}{\out{<div class="sourceCode">}}\preformatted{- / (dir)
  - a (dir)
    - e (dir)
      - i (file, size=584)
    - f (file, size=29116)
    - g (file, size=2557)
    - h.lst (file, size=62596)
  - b.txt (file, size=14848514)
  - c.dat (file, size=8504156)
  - d (dir)
    - j (file, size=4060174)
    - d.log (file, size=8033020)
    - d.ext (file, size=5626152)
    - k (file, size=7214296)
}\if{html}{\out{</div>}}

Here, there are four directories: \code{/} (the outermost directory), \code{a} and
\code{d} (which are in \code{/}), and \code{e} (which is in \code{a}). These directories
also contain files of various sizes.

Since the disk is full, your first step should probably be to find
directories that are good candidates for deletion. To do this, you need
to determine the \emph{total size} of each directory. The total size of a
directory is the sum of the sizes of the files it contains, directly or
indirectly. (Directories themselves do not count as having any intrinsic
size.)

The total sizes of the directories above can be found as follows:
\itemize{
\item The total size of directory \code{e} is \emph{584} because it contains a
single file \code{i} of size 584 and no other directories.
\item The directory \code{a} has total size \emph{94853} because it contains files
\code{f} (size 29116), \code{g} (size 2557), and \code{h.lst} (size 62596), plus
file \code{i} indirectly (\code{a} contains \code{e} which contains \code{i}).
\item Directory \code{d} has total size \emph{24933642}.
\item As the outermost directory, \code{/} contains every file. Its total size
is \emph{48381165}, the sum of the size of every file.
}

To begin, find all of the directories with a total size of \emph{at most
100000}, then calculate the sum of their total sizes. In the example
above, these directories are \code{a} and \code{e}; the sum of their total sizes
is \emph{\code{95437}} (94853 + 584). (As in this example, this process can count
files more than once!)

Find all of the directories with a total size of at most 100000. \emph{What
is the sum of the total sizes of those directories?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f07a(example_data_07())
f07b()
}
